Convert handlers to typescript
