+++ v3/mathjax3-ts/handlers/html.ts
@@ -1,5 +1,5 @@
 import {MathJax} from "../mathjax.js";
-import {HTMLHandler} from "./html/HTMLHandler.js";
+import {HTMLHandler} from './html/HTMLHandler.js';
 
 const HTML = new HTMLHandler("HTML");
 MathJax.handlers.Register(HTML);
